+++ ChangeLog.md
@@ -6,3 +6,7 @@
 
 # Commit 3 [21 Фебраля 2019]
 - PureComponents and Memoized
+
+# v1.1.0 [22 апреля 2019]
+- Adding edit status in Queue settings
+- Persistence queue settings